requirements.txt
PyQt6
pyyaml
aprslib
